+++ src/usecase/authentication/authServiceInterface.ts
@@ -1,7 +1,7 @@
 import {
   AuthServiceLoginInputDto,
   AuthServiceLoginOutputDto,
-} from '@/usecase/authentication/dto/auth.service.login.dto';
+} from '@/usecase/authentication/dto/authServiceLoginDto';
 
 export const AUTH_SERVICE_PROVIDER = 'AUTH_SERVICE_PROVIDER';
