+++ src/usecase/queries/user/dto/findByEmailDto.ts
