+++ src/usecase/queries/user/userQueryServiceInterface.ts
@@ -1,7 +1,7 @@
 import {
   FindByEmailInputDto,
   FindByEmailOutputDto,
-} from '@/usecase/queries/user/dto/find.by.email.dto';
+} from '@/usecase/queries/user/dto/findByEmailDto';
 
 export const USER_QUERY_SERVICE_PROVIDER = 'USER_QUERY_SERVICE_PROVIDER';
