+++ test/api/auth/login.e2e-spec.ts
@@ -1,11 +1,11 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { HttpStatus, INestApplication } from '@nestjs/common';
 import request from 'supertest';
-import { AppModule } from '@/app.module';
-import { AuthLoginRequest } from '@/adaptor/primary/api/auth/requests/auth.login.request.dto';
-import { AuthLoginResponse } from '@/adaptor/primary/api/auth/responses/auth.login.response.dto';
-import { createTestResponse } from '@test/common/create.test.response';
-import { prisma } from '@test/common/prisma.client';
+import { AppModule } from '@/appModule';
+import { AuthLoginRequest } from '@/adaptor/primary/api/auth/requests/authLoginRequestDto';
+import { AuthLoginResponse } from '@/adaptor/primary/api/auth/responses/authLoginResponseDto';
+import { createTestResponse } from '@test/common/createTestResponse';
+import { prisma } from '@test/common/prismaClient';
 import { User } from '@/domain/user/user';
 
 describe('【e2eテスト】/auth/login', () => {
