--- test/common/prisma.client.ts
@@ -1,3 +1,0 @@
-import { PrismaService } from '@/adaptor/primary/rdbms/prisma/prisma.service';
-
-export const prisma = new PrismaService();
