+++ test/common/prismaClient.ts
@@ -1,0 +1,3 @@
+import { PrismaService } from '@/adaptor/primary/rdbms/prisma/prismaService';
+
+export const prisma = new PrismaService();
